+++ main.py
@@ -9,7 +9,7 @@ from update import auto_update
 def main():
     logger = get_logger()
 
-    version = '0.0.1'
+    version = '0.1.0'
 
     ascii_art = f'''